dearhr/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
